OSM Geoservice plugin improvements, including offline data support
This patch contains improvements for the OSM geoservice plugin, in form
of property-controllable disk cache, and the support for a directory
containing an offline dataset.

The offline directory can contain subdirectories also in form of
symlinks. If duplicate tiles are present, the newest is used.

The population of the offline tiles registry is offloaded to a
thread to prevent blocking the startup of the application.
No synchronization is used (except on cache destruction),
so that, as the tiles are being indexed, they can be used in
the application.

For the time being, no support is given to populate such directory
(mostly due to current API limitations), but tiles can be placed
manually, for example copying them from the disk cache directory.

Task-number: QTBUG-45284
Change-Id: I518980669a3ee474545025adf05adc69cdd29781
Reviewed-by: Paolo Angelelli <paolo.angelelli@theqtcompany.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>