Enable dynamic addition/removal of map types by the mapping manager
Map types are currently fixed in the constructor, regardless of whether
they were available, or, now, whether they are enabled or not.

This patch makes the geomappingmanager notify (e.g., emit a signal)
when the available map types change at runtime. This is used in the
OSM mappingmanagerengine, which can now disable map types if provider
records turn out to be invalid or disabled after they have been
fetched.

Change-Id: I8e0e75504c882609f91c6d1ceb88424eee656f26
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>