Add support for QGeoPath
This patch adds an addition qgeoshapes, QGeoPath.
This represents a geographical path where each consecutive geopoints
are connected along the shortest line of constant bearing
(rhumb line).

The path has a width, in meters, that is used in the contains()
method to decide whether a coordinate is or not on the path, based
on shortest distance to the rhumb line segments.

Change-Id: Ia02780e3c8ac6c6d63a6083f53ea0677f8a21a1d
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>