Remove QtSystemInfo dependency from QtLocation
This was done to detect a Chinese network operator. In such cases a
different places and geocoding manager is required which takes into
account the shifting of map coordinates.

This requirement still exists in China but QtSystemInfo is not even a
released API. It makes no sense to keep the depending symbols around.
QGeoUriProvider will remain though in case Chinese customers
may want to adjust this behavior.

Change-Id: I5ee0edbf00cd561923457b449c5c9de675aa41d1
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>