Add OSRMv5 support in osm plugin
This patch adds support for the OSRMv5 protocol (the only one
currently available on the OSRM demo server
router.project.osrm.org).
Support for the old v4 protocol is retained, in case of users
who supply their own osm.routing.host parameter pointing at
a self hosted OSRMv4 server.

The support for OSRMv4 and v5 has been moved away from the
osm plugin and into the location module because this should
be used in the mapbox plugin in a subsequent patch (the routing
support in the mapbox plugin currently uses the text
direction coming from the server, which does not support i18n)

Task-number: QTBUG-56119
Change-Id: Id30fd536c7fd434011795f643221f55becfc9e18
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>