Fix panActive state when last touch point is released
When a touch interaction ends, the map would only consider the number of
points, not their state. Thus when pan was active, it would always stay
active, because in the release event, the number of points is still at
least one.

This also requires a reset of the synthetic mouse point, since otherwise the mouse
point will be copied into the list of touch points, preventing the
number of points to ever go down to zero.

Change-Id: I9821d09d75f883d01eb38b741e2b5658036be334
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>