Remove a bunch of ungrabTouch/Mouse calls
Ungrabbing means actively not wanting to receive any future events.
With slight changes in the event delivery in Qt 5.8, this leads to
release events not being delivered to the items, which in turn means
that the map thinks it's pressed after the last touch is gone.

There is no reason to call ungrab at all, when an event ends in a
regular way (release/cancel). The only reason to ungrab is when for some
reason future move/release events are unwanted, in which case the item
needs to make sure it's in the right state (as in not expecting release
events to follow).

Task-number: QTBUG-56213
Change-Id: Ieb8dbf4d8ab312cff461c4fe8a3621af67a132b9
Reviewed-by: Paolo Angelelli <paolo.angelelli@theqtcompany.com>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>