Improve support for unitary tile cost in QGeoFileTileCache
This patch adds support for separate caching mode for each of the
three caches, disk, memory and texture.

It also adds separate plugin parameters to the mapbox plugin to
control the functioning of each of these cache.

By default it only sets the disk cache to unitary, to comply with
mapbox TOS, and leaves the other two caches to bytesize behavior
(old beavior).

Change-Id: I665d40568f4e69a836d7489e1daa2d8dcb8987af
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>