Skip failing autotests due to flick signals not being emitted
It seems that the combination of mouseClick/mouseMove/mouseRelease
is not able anymore to produce consistent flicking.

This makes some autotests fail, however manually using a mouse succeed
to make the flickable flick.
For this reason this patch skips the failing autotests for the moment
until the root cause is found

Task-number: QTBUG-56679
Change-Id: If5b24ae50eca5e7f1e4d8f2d38bc7e4183c60903
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>