Position map items on their geoLeftBound
Currently map items are positioned based on their "first" coordinate.
This was because the leftbound was updated at every point update.

Now that the geo left bound is data-driven, it is possible to use it
to position elements on map. So that polylines and polygons won't
wrap around based on their first coordinate (wherever it is), but
on their geoLeftBound.
This means that two identical polygon with the coordinate list starting
at different offset will wrap around in the same way.
Same for polylines.

And circles won't wrap around anymore on their topmost point, but based on
their leftbound too.

Change-Id: I04dd04633b878f85e605af8fa84c3cd477a76e32
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>