Refactor QGeoMap API: split coordinate <-> itemPosition calls
This patch splits the two calls coordinateToItemPosition and
itemPositionToCoordinate into 3 separate steps:

- coordinate to map projection. Map projection has always to be in [0,1]
[0,1] range.
In case of QGeoTiledMap, the only map projection supported is
WebMercator, thus this call is currently equivalent to
coordinateToMercator.

- wrap map projection. Due to the longitude-wrapping-around nature of
QGeoMap/QDeclarativeGeoMap, coordinates which are greater than others
might end up before the others on the X axis of the projection.
This stage  takes care of this aspect

- finally wrapped map projection to item position. this maps a wrapped
coordinate to screen.

Equivalent calls to do the inverse conversion are also provided.

The benefits of splitting the conversion in 3 are:

- possibility to precompute the map projection of item coordinates,
thus having to simply re-wrap them (2 ifs and one sum) upon camera
change.

- Possibility to bake the last step into a 4x4 matrix and offload
this to GPU

- support for rotation and tilting cameras. Without, it would be
impossible to "unwrapBelowX" item coordinates in presence of
rotation and tilt, while now it would be possible to do this before
the final projection to screen that peforms the tilting/rotation.

Change-Id: I9b6a06ff051bbfd09e6d3584485acdcebdceb7bd
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>