Refactor: move CacheAreas to QAbstractGeoTileCache
This change moves QGeoTiledMappingManagerEngine::CacheAreas to the more
appropriate container that is QAbstractGeoTileCache, which already
uses this enum in its interface.

Together with making the virtual init() method of QAbstractGeoTileCache
public, instead of protected, this patch effectively disentangle
QAbstractGeoTileCache/QGeoFileTileCache from the
QGeoTiledMappingManagerEngine, and makes it usable potentially also
by other (tile) renderers.

Change-Id: Ic635b81660f81d86884cc6b00e3a2ab9420daa9f
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>