Add clipping support to Clip2Tri
Clip2Tri is a Poly2Tri and Clipper wrapper to do robust triangulation.
It also catches exceptions, making Clipper safe to be used with Qt.

This patch adds wrappers to expose clipping functionalities through
Clip2Tri, in order to use Clipper from a layer that sanitizes the
exception and makes it possible to use in Qt.

Change-Id: I0cf18d55c87fee2584bcf34928f1a9cb1c3e287f
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>