Prevent destroying QGeoTileProviderOsm before QGeoFileTileCacheOsm
Parenting QGeoTileProviderOsm under QGeoMappingManagerEngine puts them
at the same level in the qobject tree as QGeoFileTileCacheOsm.
Since the latter uses the providers in the destructor, it might end
up calling the destructor with providers already destroyed.

This patch parents the providers under the file tile cache, so that
they will be destroyed after the cache.

Task-number: QTBUG-58124
Change-Id: Id2fc04da58e1ba9c80a4b6c4ebeaf98cccc7d872
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>