Allow QGeoMap to render map items
Currently QtLocation Map items are always rendered by
QtLocation, on top of what QGeoMap generates.

This patch introduces a new private api call to QGeoMap,
supportedMapItemTypes(), that is used to inform QtLocation
to not render map items of those types, but rather
pass them over to the QGeoMap, which will take care of
the rendering.

In this way, more advanced renderers can properly render
map items, for example below labels or 3D buildings.

Change-Id: I1c82d4f11d4dd44c3011926512520d62e26295d4
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>