Expose QDeclarativeGeoMapCopyrightNotice in QML
This change allows the user to create a separate
MapCopyrightsNotice, and source the notice from a Map
element.

The purpose is to allow the user to stack multiple Map
elements and have a cleanly arranged set of copyright
notices on top of this stack.

To do so it is now possible to setCopyrightsVisible(false)
on all the Map elements, then declare one notice per map,
and arrange it appropriately on top of the stack.

As for any other QQuickItem, a MapCopyrightsNotice cannot
be declared inside a Map {} element, and if done so, a
a warning will be produced.

Change-Id: Ie38dec40cf25f0c44d1da39fd74469944eb79eba
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>