Add field of view support to QtLocation mapping
This patch introduces a new camera parameter, field of view, used in
QGeoCameraCapabilities, QGeoCameraData and QDeclarativeGeoMap.

This is necessary as now QtLocation supports map tilting, and different
QGeoMap implementations support different ranges for the field of view
of the camera.

Through this API it becomes possible to query the supported range and
current value, and to stack Map elements using different plugins
while keeping a consistent camera projection across all of them.

Change-Id: Ie061142bea090ec223d842efbe7924f924430496
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>