Add tilt and bearing gestures to QQuickGeoMapGestureArea
This patch adds two new gestures to the Map gesture area, that
are two finger rotation and two finger parallel vertical sliding.

The first gesture changes the bearing of the map, while the second
gesture changes the tilt angle.

The rotation gesture can co-exist with both pan and pinch.
In other words it's possible to put down two fingers on the map, and
rotate, pinch and pan at the same time.

The tilt gesture, on the other hand, excludes the others when
initiated, and also does not start, if any of the other actions is
in progress.

Change-Id: I0ef003caf0efe4addcf2e5ad563212f3c53db9ba
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>