Add color and style properties to QDeclarativeGeoMapCopyrightNotice
This patch adds two new properties to the MapCopyrightNotice element
to allow changing the background color and content style of the notice.
backgroundColor controls the color of the background of the element.
styleSheet, on the other hand, sets a stylesheet to the html,
which allows to change color, font and size of the html.

To make this work, the html output from the plugins has to be properly
enclosed in <body></body> tags.

Change-Id: I776f78f4b26b5b2cf04510a9fbc6302334b09c61
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>