Fix Mapbox GL plugin not rendering when reloaded
The code was using a 'static' flag for checking if rendering
on a separated thread, and thus, force the map refresh. Side effect
is this code path was only accessible once per application run once
the check was performed.

Make the flag a member attribute instead, so the flag is reset
on every new QGeoMapMapboxGLPrivate instance.

Change-Id: I7f2bbf2f0f21e45b9f470976cf8e7878a0f0c6f8
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>