Add mapItemOpacity signal to QDeclarativeGeoMapItemBase
Map items are different from regular QQuickItems in that they could
be rendered externally of the qtquick scene graph renderer.

With the introduction of MapItemGroups, the opacity of map items can
be changed not only by setting the opacity property of an item, but
also by changing the opacity property of a MapItemGroup, which would
affect the opacity of all the items contained inside.

When these items are handed over to a plugin for rendering, it becomes
therefore impossible to know when the opacity of the group changes.

The new signal mapItemOpacity serves the purpose of communicating
whether either the opacity property of an item, or the opacity
property of its MapItemGroup parent (if the item happens to be
inside a group) changes.

The associated mapItemOpacity is then in charge of returning the
correct combined opacity of the item.

Change-Id: I0ecbd1fc4c220291209e649bb44848854760f682
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>