Use OpenGL anisotropic texture filtering if available
Enabling 16x anisotropic texture filtering by default if available and
running on OpenGL.
This has almost no performance impact, while at the same time greatly
improving the map quality when tilting the map.

OpenGL is also supposed to be smart enough not to use it when no
anisotropy is present.

Change-Id: Ic0303b13625a28d36b258d617a9dd64f5b1a9700
Reviewed-by: Gunnar Sletta <gunnar@crimson.no>