Add a default development token for Mapbox GL styles
The idea here is to have a map working zero conf, without having
to create a token. Useful for developers that want to try Mapbox GL
maps without creating an account.

The token should not be used in production (and will display a warning
telling so). It is subject to change, throttling, etc and also a
violation of the ToS.

Also fixes runtime exception when access token is not set on
mapbox:// style.

Task-number: QTBUG-58599
Change-Id: Ie7205ccbdb3ccaee753b04116aed0f5ba35f8522
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>