Add rotation/tilt/fov sliders to the map viewer example
This patch adds a new qml element, MapSliders, that consist of a
set of sliders to control camera parameters.
This element can be hidden/shown using the included semitransparent
button.

The element can also be ported to the other examples to offer the
same controls.

This element can be positioned on the left or the right of the map,
and for this example it has been moved to the left to not occlude
the minimap.

Change-Id: Icc56d90b181c2481727e73e3d332a66c05352f57
Reviewed-by: J-P Nurmi <jpnurmi@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>