Fix incorrect handling of provider resolution for offline storage
This patch fixes the incorrect handling of the population/repopulation
of OSM offline tiles when the providers are initially populated.
In this way it is now fine if the user selects high dpi tiles, and has
both high dpi and low dpi offline tiles, and the high dpi provider
is unavailable.

Change-Id: I4d7080d6f259663bf6ca7a985fd9784fb81cfec3
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>