Evaluate the visible region lazily in QGeoProjectionWebMercator
This change optimizes QGeoProjectionWebMercator::setupCamera, factoring
out the code that calculates the visible region, and replacing it
with a dirty flag.
The computation is deferred to the first time visibleRegion() gets
called.
This is done because the visible region is currently only used to render
some types of map items, meaning that in a multi layer setup, only few
layers might need it.

Change-Id: I5dd75969581513eb60e3cf5d6ba649349bf25378
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>