Add alignCoordinateToPoint method
This method effectively extends the functionalities offered by
the center qml property, allowing to set a coordinate to an item
position other than the center of the item.
This is useful in those applications where the center of the scene
(e.g., a cursor) is not to be place exactly in the center of the map.

[ChangeLog][QtLocation][QDeclarativeGeoMap] Added alignCoordinateToPoint
method to QDeclarativeGeoMap.

Change-Id: I68a3d12e8376b83c4c1775a3af5e2125d3e71d63
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>