Fix dragging items out of map bounds
In 5.9.0 map items are clipped against the visible region.
This implies that their geometry is also clipped against
the visible region.
This is problematic in ::geometryChanged, since the old
geometry is always clipped in this way.

This patch clips items against a "projectable" region
instead, that is the part of the map that is in front
of the camera.

Since this can produce very large vertices, mapbox
earcut 3rd party library is pulled in, to replace
qTriangulate that only supports coordinates up to
1<<21.

Task-number: QTBUG-61727
Change-Id: I7449e755a4848a2b2107c5de4e27821e3e887bfb
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>