Fix places_map example view on Android
The previous setup enforced a small 360x640 view size on Android which
makes no sense on the platform's full screen premise.

This patch tells the QML rectangle to adjust to the parent's size and
shifts the size limitation to the surrounding QuickView. At the same
time the QuickView will automatically change to full screen as per
platform requirement.

Change-Id: I2dcd0fc345f6c1ef7520bb10116d9f2fe0722dff
Reviewed-by: Paolo Angelelli <paolo.angelelli@qt.io>