Fix enabling polymorphic pimpl in QGeoRoute/Segment/Maneuver
Change 8ac6377e62af803b567449cdf30c669b92114cc4 allowed to subclass
private implementations of these 3 classes.
However, the pimpl was kept in a QExplicitlySharedDataPointer for
the first two, and QSharedDataPointer for the last class.

Failing to specialize the clone method of the pointer would result in
incorrect detach behavior, which would however only be visible with the
third class (the first two never call detach() internally, and the line
QEXPECT_FAIL("", "QGeoRoute equality operators broken", Continue); in
the tests make it clear.

This change also adds the equals virtual method also to QGeoRouteSegment
and QGeoManeuver private implementations.

Change-Id: If51dafebb19e4cb2d1bcf40c2b6f136f804198e0
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>